feat(intelligence): unify Neuron agent chat history with agent_conversation_messages
Rename the lead-scoped KanvasMessageHistory to SalesAssistKanvasMessageHistory
to preserve the existing social-message-based flow used by IntelligenceCRM,
and introduce a new generic KanvasMessageHistory that persists in
intelligence.agent_conversation_messages — the same table the Laravel-AI
agents already write to via KanvasConversationStore — so Neuron and Laravel-AI
chats share a single source of truth.

Wires the new history into KanvasGenericNeuronAgent::chatHistory() using
the agent's threadId as conversation_id, and seeds threadId from the
session UUID in ProcessAgentChatAction.
