+++ src/Domains/Intelligence/Agents/Actions/ProcessAgentChatAction.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Kanvas\Intelligence\Agents\Actions;
 
+use Baka\Support\Str;
 use Kanvas\Apps\Models\Apps;
 use Kanvas\Companies\Models\Companies;
 use Kanvas\Exceptions\ValidationException;
@@ -99,6 +100,8 @@ class ProcessAgentChatAction
         }
 
         $handler->setConfiguration($this->agent, $this->session?->entity(), null, $this->user);
+        $threadId = $this->session?->uuid ?? Str::uuid();
+        $handler->setThreadId($threadId);
 
         return new RunNeuronChatAction(
             agent: $this->agent,
