+++ src/Domains/Intelligence/Agents/Neuron/KanvasGenericNeuronAgent.php
@@ -6,11 +6,29 @@ namespace Kanvas\Intelligence\Agents\Neuron;
 
 use Kanvas\NervousSystem\Capability\Models\Tool;
 use Kanvas\NervousSystem\Capability\Services\CapabilityProvider;
+use NeuronAI\Chat\History\AbstractChatHistory;
+use NeuronAI\Chat\History\InMemoryChatHistory;
 use Override;
 use ReflectionClass;
 
 class KanvasGenericNeuronAgent extends BaseKanvasAgent
 {
+    #[Override]
+    protected function chatHistory(): AbstractChatHistory
+    {
+        if ($this->user === null || $this->app === null || $this->company === null) {
+            return new InMemoryChatHistory();
+        }
+
+        return new KanvasMessageHistory(
+            app: $this->app,
+            company: $this->company,
+            user: $this->user,
+            agentClass: static::class,
+            conversationId: $this->threadId,
+        );
+    }
+
     #[Override]
     protected function tools(): array
     {
