+++ src/Domains/Intelligence/Agents/Neuron/KanvasMessageHistory.php
@@ -4,17 +4,13 @@ declare(strict_types=1);
 
 namespace Kanvas\Intelligence\Agents\Neuron;
 
-use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Str;
 use Kanvas\Apps\Models\Apps;
 use Kanvas\Companies\Models\Companies;
-use Kanvas\Social\Messages\Actions\CreateMessageAction;
-use Kanvas\Social\Messages\DataTransferObject\MessageInput;
-use Kanvas\Social\Messages\Models\AppModuleMessage;
-use Kanvas\Social\MessagesTypes\Services\MessageTypeService;
 use Kanvas\Users\Models\Users;
 use NeuronAI\Chat\Enums\MessageRole;
 use NeuronAI\Chat\History\AbstractChatHistory;
-use NeuronAI\Chat\History\ChatHistoryInterface;
 use NeuronAI\Chat\Messages\AssistantMessage;
 use NeuronAI\Chat\Messages\Message;
 use NeuronAI\Chat\Messages\UserMessage;
@@ -22,131 +18,69 @@ use Override;
 
 class KanvasMessageHistory extends AbstractChatHistory
 {
-    private const string AGENT_VERB = 'agent';
-    private const string USER_VERB = 'user';
-
-    // Verbos que nunca deben exponerse al lead
-    private const array INTERNAL_VERBS = ['notes', 'ai_assist', 'internal'];
+    private const string CONNECTION = 'intelligence';
+    private const string TABLE_CONVERSATIONS = 'agent_conversations';
+    private const string TABLE_MESSAGES = 'agent_conversation_messages';
 
     public function __construct(
         private readonly Apps $app,
         private readonly Companies $company,
         private readonly Users $user,
-        private readonly Model $entity,
-        private readonly ?string $threadId = null,
-        private readonly bool $includeInternal = false,
+        private readonly string $agentClass,
+        private ?string $conversationId = null,
         int $contextWindow = 50000,
     ) {
         parent::__construct($contextWindow);
-        $this->load();
+
+        $this->conversationId ??= $this->findLatestConversation();
+
+        if ($this->conversationId !== null) {
+            $this->load();
+        }
+    }
+
+    public function getConversationId(): ?string
+    {
+        return $this->conversationId;
+    }
+
+    private function findLatestConversation(): ?string
+    {
+        return DB::connection(self::CONNECTION)
+            ->table(self::TABLE_CONVERSATIONS)
+            ->where('user_id', $this->user->getId())
+            ->where('apps_id', $this->app->getId())
+            ->where('companies_id', $this->company->getId())
+            ->orderBy('updated_at', 'desc')
+            ->first()?->id;
     }
 
     private function load(): void
     {
-        $messages = AppModuleMessage::query()
-            ->where('system_modules', get_class($this->entity))
-            ->where('entity_id', $this->entity->getKey())
-            ->where('apps_id', $this->app->getId())
-            ->whereHas('message', fn ($q) => $q->where('is_deleted', 0))
-            ->with(['message.messageType'])
+        $messages = DB::connection(self::CONNECTION)
+            ->table(self::TABLE_MESSAGES)
+            ->where('conversation_id', $this->conversationId)
+            ->orderBy('created_at', 'asc')
             ->orderBy('id', 'asc')
             ->get()
-            ->map(function (AppModuleMessage $appModuleMessage): ?Message {
-                $socialMessage = $appModuleMessage->message;
+            ->map(function ($row): ?Message {
+                $content = (string) ($row->content ?? '');
 
-                if (! $socialMessage) {
+                if ($content === '') {
                     return null;
                 }
 
-                $stored = $socialMessage->getMessage();
-
-                if ($this->threadId !== null && ($stored['thread_id'] ?? null) !== $this->threadId) {
-                    return null;
-                }
-
-                $verb = $socialMessage->messageType?->verb ?? self::USER_VERB;
-
-                if (! $this->includeInternal && in_array($verb, self::INTERNAL_VERBS, true)) {
-                    return null;
-                }
-
-                $text = (string) ($stored['content'] ?? $stored['text'] ?? '');
-
-                if ($text === '') {
-                    return null;
-                }
-
-                $fromIa = (bool) ($stored['from_ia'] ?? false);
-                $fromHuman = (bool) ($stored['from_human'] ?? false);
-
-                $channel = $socialMessage->channels()->first();
-                $isInternal = $channel?->isNoteChannel() || $channel?->isAiAssistChannel();
-
-                if ($isInternal) {
-                    $prefixed = "[INTERNAL - {$channel->name}] {$text}";
-                } elseif ($fromIa) {
-                    $prefixed = "[Assistant] {$text}";
-                } elseif ($fromHuman) {
-                    $owner = $socialMessage->user?->displayname ?: 'Owner';
-                    $prefixed = "[Owner - {$owner}] {$text}";
-                } else {
-                    $prefixed = '[' . $this->entityIdentityLabel() . "] {$text}";
-                }
-
-                return $fromIa
-                    ? new AssistantMessage($prefixed)
-                    : new UserMessage($prefixed);
+                return $row->role === MessageRole::ASSISTANT->value
+                    ? new AssistantMessage($content)
+                    : new UserMessage($content);
             })
             ->filter()
             ->values()
-            ->toArray();
+            ->all();
 
-        $coalesced = [];
-        foreach ($messages as $m) {
-            $last = end($coalesced) ?: null;
-            if ($last !== null && $last->getRole() === $m->getRole()) {
-                $last->setContents($last->getContent() . "\n\n" . $m->getContent());
-
-                continue;
-            }
-            $coalesced[] = $m;
+        if (! empty($messages)) {
+            $this->history = $messages;
         }
-
-        while (! empty($coalesced) && $coalesced[0]->getRole() !== MessageRole::USER->value) {
-            array_shift($coalesced);
-        }
-
-        if (! empty($coalesced)) {
-            $this->history = array_values($coalesced);
-        }
-    }
-
-    #[Override]
-    public function addMessage(Message $message): ChatHistoryInterface
-    {
-        $last = end($this->history) ?: null;
-        if ($last !== null && $last->getRole() === $message->getRole()) {
-            $last->setContents($last->getContent() . "\n\n" . $message->getContent());
-            $this->trimHistory();
-            $this->onNewMessage($message);
-            $this->setMessages($this->history);
-
-            return $this;
-        }
-
-        return parent::addMessage($message);
-    }
-
-    private function entityIdentityLabel(): string
-    {
-        if (method_exists($this->entity, 'people') && $this->entity->people) {
-            $name = (string) $this->entity->people->getName();
-            if ($name !== '') {
-                return "Lead - {$name}";
-            }
-        }
-
-        return class_basename($this->entity) . ':' . $this->entity->getKey();
     }
 
     #[Override]
@@ -164,33 +98,46 @@ class KanvasMessageHistory extends AbstractChatHistory
             return;
         }
 
-        $isAssistant = $role === MessageRole::ASSISTANT->value;
-        $verb = $isAssistant ? self::AGENT_VERB : self::USER_VERB;
-        $messageType = MessageTypeService::getOrCreate($this->app, $verb);
-
-        $messageData = [
-            'content' => $content,
-            'from_me' => $isAssistant,
-            'from_ia' => $isAssistant,
-            'from_human' => ! $isAssistant,
-        ];
-
-        if ($this->threadId !== null) {
-            $messageData['thread_id'] = $this->threadId;
+        if ($this->conversationId === null) {
+            $this->conversationId = $this->createConversation($content);
+        } else {
+            DB::connection(self::CONNECTION)
+                ->table(self::TABLE_CONVERSATIONS)
+                ->where('id', $this->conversationId)
+                ->update(['updated_at' => now()]);
         }
 
-        $createMessageAction = new CreateMessageAction(
-            new MessageInput(
-                app: $this->app,
-                company: $this->company,
-                user: $this->user,
-                type: $messageType,
-                message: $messageData,
-                is_public: 0,
-            )
-        );
-        $createMessageAction->runWorkflow = false;
-        $socialMessage = $createMessageAction->execute();
-        $socialMessage->addEntity($this->entity);
+        DB::connection(self::CONNECTION)->table(self::TABLE_MESSAGES)->insert([
+            'id' => (string) Str::uuid7(),
+            'conversation_id' => $this->conversationId,
+            'user_id' => $this->user->getId(),
+            'agent' => $this->agentClass,
+            'role' => $role,
+            'content' => $content,
+            'attachments' => '[]',
+            'tool_calls' => '[]',
+            'tool_results' => '[]',
+            'usage' => '[]',
+            'meta' => '[]',
+            'created_at' => now(),
+            'updated_at' => now(),
+        ]);
+    }
+
+    private function createConversation(string $firstMessage): string
+    {
+        $conversationId = (string) Str::uuid7();
+
+        DB::connection(self::CONNECTION)->table(self::TABLE_CONVERSATIONS)->insert([
+            'id' => $conversationId,
+            'user_id' => $this->user->getId(),
+            'apps_id' => $this->app->getId(),
+            'companies_id' => $this->company->getId(),
+            'title' => Str::limit($firstMessage, 100, ''),
+            'created_at' => now(),
+            'updated_at' => now(),
+        ]);
+
+        return $conversationId;
     }
 }
